refactor(lowering): plan := mem transfers before emission (#916)
- Add ldTransferPlan.ts with planAssignmentMemTransfer and scratch/preservation
  decisions for address-of, byte mem->mem, and word mem->mem := paths.
- ldEncoding emitAssignmentMemTransfer runs the plan via runAssignmentMemTransferPlan;
  emitLdForm stays the emission entry; recursive emitLdForm for staged subforms unchanged.
- Remove inlined register-usage and direct-path helpers from ldEncoding (moved to planner).

Made-with: Cursor
